requirements.txt
flask
boto3
gunicorn
requests
python-dotenv